Add async metric writing API to match update check API.
